Remove UI class,  I/O redirection testing, reorganise exception classes
