+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,6 +1,5 @@
 package duke.commands;
 
-import duke.core.Ui;
 import duke.exceptions.MeanDukeException;
 
 /**
