--- text-ui-test/input.txt
@@ -1,33 +1,0 @@
-hi
-add
-add task
-add todo
-add todo todo1
-add deadline
-add deadline deadline1
-add deadline deadline1 /by
-add deadline deadline1 /by 1st jan
-add deadline deadline1 /by 2023-12-12
-add deadline deadline1 /by 2023-12-12 99:99
-add deadline deadline1 /by 2023-12-12 23:43
-add event
-add event event1
-add event event1 /from /to
-add event event1 /from 2023-12-12 /to 1st jan
-add event event1 /from 1st jan /to 2023-12-12
-add event event1 /from 2023-12-12 /to 2023-12-12
-add event event1 /from 2023-12-12 99:99 /to 2023-12-12 23:59
-add event event1 /from 2023-12-12 23:59 /to 2023-12-12 99:99
-add event event1 /from 2023-12-12 23:59 /to 2023-12-12 23:59
-mark
-mark 0
-mark 1
-mark 1
-unmark
-unmark 0
-unmark 1
-unmark 1
-delete
-delete 0
-delete 1
-end
